Crate cryptopals_rust

rust-cryptopals

Rust implementation of the cryptopals set of challenges.

Includes a simple CLI tool to try out the implementations.

For the time being, I’m not going to work on more challenges. I’m currently getting more familliar with the Rust ecosystem, and want to easily be able to apply new concepts to a small project as I move forward. Once I’m through, I’ll work more actively on more of these challenges.

https://cryptopals.com/

To do

set 1

challenge 1

challenge 2

challenge 3

challenge 4

challenge 5

challenge 6

challenge 7

challenge 8

set 2

challenge 9

challenge 10

challenge 11

challenge 12

challenge 13

challenge 14

challenge 15

challenge 16

set 3

challenge 17

challenge 18

challenge 19

challenge 20

challenge 21

challenge 22

challenge 23

challenge 24

set 4

challenge 25

challenge 26

challenge 27

challenge 28

challenge 29

challenge 30

challenge 31

challenge 32

set 5

challenge 33

challenge 34

challenge 35

challenge 36

challenge 37

challenge 38

challenge 39

challenge 40

set 6

challenge 41

challenge 42

challenge 43

challenge 44

challenge 45

challenge 46

challenge 47

challenge 48

set 7

challenge 49

challenge 50

challenge 51

challenge 52

challenge 53

challenge 54

challenge 55

challenge 56

set 8

challenge 57

challenge 58

challenge 59

challenge 60

challenge 61

challenge 62

challenge 63

challenge 64

Modules

set1

Structs

Bytes

This newtype will make it easier to handle bytes for cryptographic operations. A newtype is preferable to an alias, to avoid having bytes-specific methods apply too freely with other Vec-based data types, and vice-versa. e.g. a method for Bytes will not work for instance on a type Ascii = Vec<u8>, which is a good thing : )

EncodingError

Generic error structure for all encoding errors

Enums

EncodingType